Added prototype arduino sketch
This uses the proximity sensor with an Arduino. Gives us the distance over a serial, and lights up the LEDS based on the distance. Now we need to figure out how to use mutliple sensors at the same time, use an LED strip, and really understand how distance is calculated with the Trig and Echo Signals.